hlavne menu

+++ MacroMeter/User.cs
@@ -7,10 +7,10 @@
         public string Email { get; set; }
         public string Password { get; set; }
         public double Kalorie { get; set; }  
-        public int CielovaVaha { get; set; }
+        public double CielovaVaha { get; set; }
 
         public int Vek { get; set; }
-        public int Vaha { get; set; }
+        public double Vaha { get; set; }
         public int Vyska { get; set; }
         public string Pohlavie { get; set; }
         public string Aktivita { get; set; }
